fix: use golang:1.21.0
