+++ Dockerfile
@@ -1,5 +1,5 @@
 # UI build stage
-FROM node:16-alpine3.14 as js-builder
+FROM node:16-alpine3.17 as js-builder
 
 ENV NODE_OPTIONS=--max_old_space_size=8000
 WORKDIR /clickvisual
@@ -12,7 +12,7 @@ RUN yarn build
 
 
 # API build stage
-FROM golang:1.18.6-alpine3.16 as go-builder
+FROM golang:1.21.0-alpine3.17 as go-builder
 ARG GOPROXY=goproxy.cn
 
 ENV GOPROXY=https://${GOPROXY},direct
@@ -32,7 +32,7 @@ RUN ls -rlt ./api/internal/ui/dist && make build.api
 
 
 # Fianl running stage
-FROM alpine:3.14.3
+FROM alpine:3.17
 LABEL maintainer="clickvisual@shimo.im"
 
 WORKDIR /clickvisual
