+++ api/core/i/switcher.go
@@ -1,10 +1,10 @@
-package ifstorer
+package i
 
 import (
 	"database/sql"
 )
 
-type Storer interface {
+type Switcher interface {
 	// Description read model
 	Description() string
 
@@ -14,7 +14,7 @@ type Storer interface {
 	Attach() error
 }
 
-type Params struct {
+type SwitcherParams struct {
 	// common
 	CreateType int
 
@@ -25,10 +25,13 @@ type Params struct {
 	Table     string  // table name
 	Conn      *sql.DB // clickhouse
 
-	// storer
-	Fields string
-	TTL    int // ttl Data expiration time, unit is the day
-
 	// switcher
-
+	RawLogField         string
+	RawLogFieldParent   string
+	ParseIndexes        string
+	ParseFields         string
+	ParseTime           string
+	ParseWhere          string
+	IsRawLogFieldString bool
+	CustomTimeField     string
 }
