+++ api/core/storer/storer.go
@@ -1,12 +1,12 @@
 package storer
 
 import (
+	"github.com/clickvisual/clickvisual/api/core/i"
 	"github.com/clickvisual/clickvisual/api/core/storer/clickhouse"
-	"github.com/clickvisual/clickvisual/api/core/storer/ifstorer"
 	"github.com/clickvisual/clickvisual/api/pkg/model/db"
 )
 
-func New(ds string, params ifstorer.Params) ifstorer.Storer {
+func New(ds string, params i.StorerParams) i.Storer {
 	switch ds {
 	case db.DatasourceClickHouse:
 		return clickhouse.NewStorer(params)
