+++ api/core/switcher/switcher.go
@@ -1,12 +1,12 @@
 package switcher
 
 import (
+	"github.com/clickvisual/clickvisual/api/core/i"
 	"github.com/clickvisual/clickvisual/api/core/switcher/clickhouse"
-	"github.com/clickvisual/clickvisual/api/core/switcher/ifswitcher"
 	"github.com/clickvisual/clickvisual/api/pkg/model/db"
 )
 
-func New(ds string, params ifswitcher.Params) ifswitcher.Switcher {
+func New(ds string, params i.SwitcherParams) i.Switcher {
 	switch ds {
 	case db.DatasourceClickHouse:
 		return clickhouse.NewSwitcher(params)
